fix: add progress notifs to torrent

+++ Code/shared/events/progress.cpp
@@ -1,10 +1,10 @@
 #include <progress.hpp>
 
-ProgressUpdateMessage::ProgressUpdateMessage(float bps, float progress) : Event(EventType::kOnUpdate) {
+ProgressUpdateMessage::ProgressUpdateMessage(long double bps, long double progress) : Event(EventType::kOnUpdate) {
   m_fBps = bps;
   m_fProgress = progress;
 }
 
-float ProgressUpdateMessage::GetBps() { return m_fBps; }
+long double ProgressUpdateMessage::GetBps() { return m_fBps; }
 
-float ProgressUpdateMessage::GetProgress() { return m_fProgress; }
+long double ProgressUpdateMessage::GetProgress() { return m_fProgress; }
